Implement OverlaySortTabControl component
