+++ osu.Game/Overlays/Comments/CommentsHeader.cs
@@ -16,8 +16,6 @@ namespace osu.Game.Overlays.Comments
 {
     public class CommentsHeader : CompositeDrawable
     {
-        private const int font_size = 14;
-
         public readonly Bindable<CommentsSortCriteria> Sort = new Bindable<CommentsSortCriteria>();
         public readonly BindableBool ShowDeleted = new BindableBool();
 
@@ -40,29 +38,11 @@ namespace osu.Game.Overlays.Comments
                     Padding = new MarginPadding { Horizontal = 50 },
                     Children = new Drawable[]
                     {
-                        new FillFlowContainer
+                        new OverlaySortTabControl<CommentsSortCriteria>
                         {
-                            AutoSizeAxes = Axes.Both,
-                            Direction = FillDirection.Horizontal,
-                            Spacing = new Vector2(10, 0),
                             Anchor = Anchor.CentreLeft,
                             Origin = Anchor.CentreLeft,
-                            Children = new Drawable[]
-                            {
-                                new OsuSpriteText
-                                {
-                                    Anchor = Anchor.CentreLeft,
-                                    Origin = Anchor.CentreLeft,
-                                    Font = OsuFont.GetFont(size: font_size),
-                                    Text = @"Sort by"
-                                },
-                                new SortTabControl
-                                {
-                                    Anchor = Anchor.CentreLeft,
-                                    Origin = Anchor.CentreLeft,
-                                    Current = Sort
-                                }
-                            }
+                            Current = Sort
                         },
                         new ShowDeletedButton
                         {
@@ -106,7 +86,7 @@ namespace osu.Game.Overlays.Comments
                         {
                             Anchor = Anchor.CentreLeft,
                             Origin = Anchor.CentreLeft,
-                            Font = OsuFont.GetFont(size: font_size),
+                            Font = OsuFont.GetFont(size: 12),
                             Text = @"Show deleted"
                         }
                     },
@@ -125,5 +105,12 @@ namespace osu.Game.Overlays.Comments
                 return true;
             }
         }
+    }
+
+    public enum CommentsSortCriteria
+    {
+        New,
+        Old,
+        Top
     }
 }
